# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bakaya MSME AR Desk API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Buyer {
  id: number;
  name: string;
  contactName: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  /** Preferred language code e.g. 'ta', 'hi', 'en' */
  language: string;
  /** @nullable */
  gstNumber?: string | null;
  /** @nullable */
  city?: string | null;
  totalOutstanding?: number;
  invoiceCount?: number;
  createdAt: string;
}

export interface BuyerInput {
  name: string;
  contactName: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  language: string;
  /** @nullable */
  gstNumber?: string | null;
  /** @nullable */
  city?: string | null;
}

export interface BuyerUpdate {
  name?: string;
  contactName?: string;
  phone?: string;
  /** @nullable */
  email?: string | null;
  language?: string;
  /** @nullable */
  gstNumber?: string | null;
  /** @nullable */
  city?: string | null;
}

export type InvoiceStatus = typeof InvoiceStatus[keyof typeof InvoiceStatus];


export const InvoiceStatus = {
  pending: 'pending',
  overdue: 'overdue',
  escalating: 'escalating',
  paid: 'paid',
  disputed: 'disputed',
} as const;

export type InvoiceEscalationStage = typeof InvoiceEscalationStage[keyof typeof InvoiceEscalationStage];


export const InvoiceEscalationStage = {
  none: 'none',
  nudge: 'nudge',
  tax_nudge: 'tax_nudge',
  formal_demand: 'formal_demand',
  odr_ready: 'odr_ready',
} as const;

export interface Invoice {
  id: number;
  invoiceNumber: string;
  buyerId: number;
  buyerName: string;
  amount: number;
  invoiceDate: string;
  dueDate: string;
  status: InvoiceStatus;
  escalationStage: InvoiceEscalationStage;
  daysOverdue: number;
  interestAccrued?: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  poNumber?: string | null;
  createdAt: string;
}

export type InvoiceDetailStatus = typeof InvoiceDetailStatus[keyof typeof InvoiceDetailStatus];


export const InvoiceDetailStatus = {
  pending: 'pending',
  overdue: 'overdue',
  escalating: 'escalating',
  paid: 'paid',
  disputed: 'disputed',
} as const;

export type InvoiceDetailEscalationStage = typeof InvoiceDetailEscalationStage[keyof typeof InvoiceDetailEscalationStage];


export const InvoiceDetailEscalationStage = {
  none: 'none',
  nudge: 'nudge',
  tax_nudge: 'tax_nudge',
  formal_demand: 'formal_demand',
  odr_ready: 'odr_ready',
} as const;

export type EscalationEventStage = typeof EscalationEventStage[keyof typeof EscalationEventStage];


export const EscalationEventStage = {
  nudge: 'nudge',
  tax_nudge: 'tax_nudge',
  formal_demand: 'formal_demand',
  odr_ready: 'odr_ready',
} as const;

export type EscalationEventChannel = typeof EscalationEventChannel[keyof typeof EscalationEventChannel];


export const EscalationEventChannel = {
  whatsapp: 'whatsapp',
  email: 'email',
  system: 'system',
} as const;

export interface EscalationEvent {
  id: number;
  invoiceId: number;
  stage: EscalationEventStage;
  message: string;
  sentAt: string;
  channel: EscalationEventChannel;
  approvedByOwner: boolean;
  /** @nullable */
  language?: string | null;
}

export interface InvoiceDetail {
  id: number;
  invoiceNumber: string;
  buyerId: number;
  buyerName: string;
  amount: number;
  invoiceDate: string;
  dueDate: string;
  status: InvoiceDetailStatus;
  escalationStage: InvoiceDetailEscalationStage;
  daysOverdue: number;
  interestAccrued?: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  poNumber?: string | null;
  createdAt: string;
  escalationEvents: EscalationEvent[];
}

export interface InvoiceInput {
  invoiceNumber: string;
  buyerId: number;
  amount: number;
  invoiceDate: string;
  dueDate: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  poNumber?: string | null;
}

export type InvoiceUpdateStatus = typeof InvoiceUpdateStatus[keyof typeof InvoiceUpdateStatus];


export const InvoiceUpdateStatus = {
  pending: 'pending',
  overdue: 'overdue',
  escalating: 'escalating',
  paid: 'paid',
  disputed: 'disputed',
} as const;

export interface InvoiceUpdate {
  amount?: number;
  dueDate?: string;
  status?: InvoiceUpdateStatus;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  poNumber?: string | null;
}

export type EscalationInputStage = typeof EscalationInputStage[keyof typeof EscalationInputStage];


export const EscalationInputStage = {
  nudge: 'nudge',
  tax_nudge: 'tax_nudge',
  formal_demand: 'formal_demand',
  odr_ready: 'odr_ready',
} as const;

export type EscalationInputChannel = typeof EscalationInputChannel[keyof typeof EscalationInputChannel];


export const EscalationInputChannel = {
  whatsapp: 'whatsapp',
  email: 'email',
  system: 'system',
} as const;

export interface EscalationInput {
  stage: EscalationInputStage;
  channel: EscalationInputChannel;
  approvedByOwner: boolean;
  /** @nullable */
  customMessage?: string | null;
}

export interface InterestCalc {
  invoiceId: number;
  principalAmount: number;
  daysOverdue: number;
  rbiRate: number;
  /** 3x RBI rate as per MSMED Act */
  applicableRate: number;
  dailyInterest: number;
  totalInterest: number;
  totalDue: number;
  isLegallyOverdue: boolean;
  section43bhApplies?: boolean;
}

export interface DashboardSummary {
  totalInvoices: number;
  totalOutstanding: number;
  overdueCount: number;
  overdueAmount: number;
  escalatingCount: number;
  paidCount: number;
  recoveredAmount: number;
  avgDaysOverdue: number;
  odrReadyCount?: number;
}

export interface OverdueStage {
  stage: string;
  count: number;
  amount: number;
  label?: string;
}

export interface ActivityItem {
  id: number;
  invoiceId: number;
  invoiceNumber: string;
  buyerName: string;
  stage: string;
  message: string;
  sentAt: string;
  channel: string;
}

export type ListInvoicesParams = {
status?: string;
buyerId?: number;
};