# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bakaya MSME AR Desk API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityItem {
  id: number;
  invoiceId: number;
  invoiceNumber: string;
  buyerName: string;
  stage: string;
  message: string;
  sentAt: Date;
  channel: string;
}