# lib/api-zod/src/generated/types/buyerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bakaya MSME AR Desk API
 * OpenAPI spec version: 0.1.0
 */

export interface BuyerInput {
  name: string;
  contactName: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  language: string;
  /** @nullable */
  gstNumber?: string | null;
  /** @nullable */
  city?: string | null;
}