# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bakaya MSME AR Desk API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalInvoices: number;
  totalOutstanding: number;
  overdueCount: number;
  overdueAmount: number;
  escalatingCount: number;
  paidCount: number;
  recoveredAmount: number;
  avgDaysOverdue: number;
  odrReadyCount?: number;
}